Renamed file to tidyAnalysis for now.
Added aggregated averages of miles and mph to the Time Table that are tied to the type of bike and membership status of that particular timeframe (down to the minute).

It didn't seem to take very long to create the table with DuckDB. I hope it will make visualizing trends and further analysis simpler. I think it will save a lot of extra code, at least.